refactor(QuizModule) : move interface file
- improve semantics interface and reusability of interfaces
