+++ src/quiz/interface/quiz-context.ts
@@ -1,0 +1,9 @@
+export interface QuizContext {
+  artist?: string;
+
+  tag?: string;
+
+  style?: string;
+
+  page: number;
+}
