+++ src/quiz/interface/quiz-status.ts
@@ -1,0 +1,7 @@
+import { QuizContext } from './quiz-context';
+
+export interface QuizStatus {
+  context: QuizContext;
+  currentIdx: number;
+  endIdx: number;
+}
